[FFM-1363]: JS SDK handles segment updates incorrectly
What:
When a patch event for a target-segment is received we fetch all flag variations
and generate a change event for flags whos variation has changed for this target.

Why:
When rules are changed the ff service will send an event to the client SDKs.
SDKs should then fetch the updated data. If flag rules change, then only that flag needs to
be fetched, but  If segment rules change, then all flag vairations must be fetched.

Previously when we received a patch event for a target segment, we tried to fetch an
invidual flag, and were passing the segment name instead of the flag name, so this resulted
in a 404.