FFM-8141 - Add refreshEvaluations() API method (#84)
* FFM-8141 - Add refreshEvaluations() API method to allow webviews inside mobile apps coming to foreground to get latest flags

What
Adds a new function called refreshEvaluations() to CfClient

Why
This change adds a new API called refreshEvaluations() which can be called by a mobile app using the SDK in a webview coming
to the foreground to update any SSE flag events that were missed while the app was suspended. Note this new API should
not be used to poll the Feature Flags server, it will only retrieve evaluations from the server if enough time has
elapsed to avoid putting to much pressure on the backend.

It's up to the native app to call this JS method from the webview APIs when the app comes to foreground.

Testing
Manual. Tested with Xcode + iOS Simulator with a WKWebView

---------

Co-authored-by: Kevin Nagurski <Kevin.nagurski@harness.io>