[FFM-9054] - An implementation of SSE using fetch()
What
Remove XMLHttpRequest() and use fetch() instead.

This is an experimential PR, some features are absent:
- SSE timeout handling
- HTTP error code handling
- Restarting SSE

Why
This is a discovery ticket to look at implementing fetch() instead of XMLHttpRequest() and to see how feasible doing that is. Fetch has a cleaner and modern API which can help with code readability.

Testing
Manual